solidity_0.8/mutli call

+++ solidity_0.8/Call.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8;
 
 contract TestCall {
     string public message;
-    uint public x;
+    uint256 public x;
 
     event Log(string message);
 
@@ -11,7 +11,11 @@ contract TestCall {
         emit Log("fallback was called");
     }
 
-    function foo(string memory _message, uint _x) external payable returns(bool, uint) {
+    function foo(string memory _message, uint256 _x)
+        external
+        payable
+        returns (bool, uint256)
+    {
         message = _message;
         x = _x;
         return (true, 999);
@@ -20,18 +24,19 @@ contract TestCall {
 
 contract Call {
     bytes public data;
+
     function callFoo(address _test) external payable {
         (bool success, bytes memory _data) = _test.call{value: 111}(
-            abi.encodeWithSignature(
-                "foo(string,uint256)", "call foo", 123
-            )
+            abi.encodeWithSignature("foo(string,uint256)", "call foo", 123)
         );
         require(success, "call failed");
         data = _data;
     }
 
     function callDoesNotExit(address _test) external {
-        (bool success, ) = _test.call(abi.encodeWithSignature("doesNotExist()"));
+        (bool success, ) = _test.call(
+            abi.encodeWithSignature("doesNotExist()")
+        );
         require(success, "call failed");
     }
 }
